requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.1
# Required by django-debug-toolbar
sqlparse==0.1.11

coverage==3.7.1
flake8==2.2.5
# Required by flake8
pyflakes==0.8.1
pep8==1.5.7
mccabe==0.3

selenium==2.44.0
factory_boy==2.4.1

# Required for deployment but don't work on Python 3
# PyYAML==3.10
# Fabric==1.8.2
# Required by Fabric
# paramiko==1.13.0
# pycrypto==2.6.1
# ecdsa==0.11
